+++ Assets/Standard Assets/Serializing/NPCData.cs
@@ -5,5 +5,5 @@ using System.Collections.Generic;
 public class NPCData {
 	[XmlAttribute("name")]
     public string name;
-    public float disposition;
+    public int disposition;
 }
